Saque los archivos del Preview content. cree la carpeta Store donde le agregue Store View Como dos y Presentation 2
